+++ vite.config.ts
@@ -1,7 +1,9 @@
 import { defineConfig } from "vite";
 
 export default defineConfig({
+  base: "./",
   build: {
+    copyPublicDir: false,
     lib: {
       entry: "src/index.ts",
       formats: ["es"],